// cocos_client2D/assets/scripts/Generated/Model/demo/protocol/messages.ts
export interface MapEntitySnapshot {
  unitId: number;
  x: number;
  y: number;
  heading: number;
  alive: boolean;
  state: Uint8Array;
  account: string;
  cellX: number;
  cellY: number;
}
// ...
export interface S2C_Login extends IResponse {
  message?: string;
  error?: number;
  rpcId?: number;
  account: string;
  service: string;
  loginCount: number;
  token: string;
  gateName: string;
  gateIp: string;
  gatePort: number;
}
// ...
export interface G2C_EnterMap extends IResponse {
  message?: string;
  error?: number;
  rpcId?: number;
  account: string;
  mapService: string;
  mapId: number;
  unitId: number;
  x: number;
  y: number;
  entities: readonly MapEntitySnapshot[];
  fixedUpdateMs: number;
}
// ...
export interface G2C_MapReady extends IMessage {
  account: string;
  mapId: number;
  unitId: number;
  x: number;
  y: number;
}